requirements.txt
tensorflow==2.11.0
keras==2.11.0
streamlit
numpy
pillow

